requirements.txt
SQLAlchemy==2.0.35
alembic==1.13.3
asgiref==3.8.1
blinker==1.8.2
click==8.1.7
Django==5.1.1
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-enum==1.5.1
packaging==24.1
PyMySQL==1.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.35
sqlparse==0.5.1
typing_extensions==4.12.2
Werkzeug==3.0.4
WTForms==3.1.2
sqlalchemy-utils==0.41.2